test(controller, instrumenter): move instrument/revert test cases
Move almost all instrument/revert workload test cases from
dash0_controller_test.go to instrumenter_test.go.
